server/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai-whisper
torch
soundfile